// src/types/relatorio/acao20xe.ts
export type Acao20XEType = {
  codigo: number;
  codAcao: string;
  nomeCodAcao: string;
  ugExec: string;
  nomeUgExec: string;
  grupoDespesa: number;
  nomeGrupoDespesa: string;
  natDespesa: number;
  nomeNatDespesa: string;
  elementoDespesa: string;
  nomeElementoDespesa: string;
  provisaoRecebida: number;
  creditoDisponivel: number;
  despesaEmpenhada: number;
  despesaLiquidada: number;
  despesaPaga: number;
};